US179 Task 198.  Taking some of the logic out of the MedicationCapturePage but importing it from another js file
